% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_helpers.R
\name{isotopologues}
\alias{isotopologues}
\title{Calculate isotopologues given a metabolite from metabolite_dt, an ion, and the atom for which isotopes have been substituted.}
\usage{
isotopologues(search, .atom = "C", .ion = "-H", maxn = 100)
}
\arguments{
\item{search}{Character. Name or abbreviation string.  Case sensitive.}

\item{.atom}{Character. Name of the desired atom's isotope masses. Must be defined in metabolite_shifts}

\item{.ion}{Character. Name of the desired ion transformation. Must be defined in metabolite_ions}

\item{maxn}{Numeric. The largest isotopologue desired}
}
\value{
m/z vector corresponding to the supplied formula and ion and shift.
}
\description{
\code{m} Searches metabolite_dt in environment pleiad.env for rows which match supplied search either by name or abbreviation and returns the mass corresponding to that formula.
}
