% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_helpers.R
\name{mass}
\alias{mass}
\title{Retrieve a neutral mass from the metabolite_dt}
\usage{
mass(search)
}
\arguments{
\item{search}{Character. Name or abbreviation string.  Case sensitive.}
}
\value{
mass corresponding to the supplied formula
}
\description{
\code{m} Searches metabolite_dt in environment pleiad.env for rows which match supplied search either by name or abbreviation and returns the mass corresponding to that formula.
}
