% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata.R
\name{rawdata}
\alias{rawdata}
\title{Import .mzXML data into a list}
\usage{
rawdata(file, splits = NULL)
}
\arguments{
\item{file}{Character. Path to the .mzXML file}

\item{splits}{Matrix with N rows and 2 columns containing the mass ranges to extract from sequential scans.  Column 1 is the start of the mass range and column 2 is the end of the mass range.  The first row will apply to the first scan in the data file, the second to the second, and so on.  When the matrix rows are exhausted it will return to the first row. For example, a two row matrix, the first row will apply to all odd scans and the second to all even scans. Each set of N scans will be combined into a single scan for further analysis.}
}
\value{
List containing lists named k (mass peaks), s (scan information), and metadata.
}
\description{
\code{rawdata} reads a single .mzXML file into a list object which consists of a named list of data.table objects. These lists contain the raw mass spectral data from the file.
}
\details{
Also handles scan stitching if multiple scan segments were employed.
}
\examples{
\dontrun{
rawdata('file1.mzXML', splits = rbind(c(0, 300), c(300, 3000)))
}

}
