% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_helpers.R
\name{retrieve.metabolite_dt}
\alias{retrieve.metabolite_dt}
\title{Retrieve a single column from the metabolite_dt}
\usage{
retrieve.metabolite_dt(search, col = "formula")
}
\arguments{
\item{search}{Character. Name or abbreviation string.  Case sensitive.}

\item{col}{Character. Name of the desired column.}
}
\value{
value from metabolite_dt
}
\description{
\code{m} Searches metabolite_dt in environment pleiad.env for rows which match supplied search either by name or abbreviation and returns the specified col.
}
