% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspec.R
\name{subspec_masscorrect}
\alias{subspec_masscorrect}
\title{Correct deviation from target masses within subspectra}
\usage{
subspec_masscorrect(specs, maxppm = Inf)
}
\arguments{
\item{specs}{data.table containing columns mz, subspec}

\item{maxppm}{the maximum ppm error for an observation to be used in calculating the correction}
}
\value{
data.table containing new column mz.observed and a corrected mz value
}
\description{
\code{subspec_masscorrect} for every subspec corrects the average mz.target - mz
}
