% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspec.R
\name{subspec_nearestmasses}
\alias{subspec_nearestmasses}
\title{Return nearest mass peaks to supplied peaks in a set of spectra.}
\usage{
subspec_nearestmasses(masses, specs)
}
\arguments{
\item{masses}{data.table containing column mz}

\item{specs}{data.table containing columns mz, s}
}
\value{
data.table containing columns mz.target, mz, ppm, s, as well as original columns.
}
\description{
\code{specs_nearestmasses} for every spectra returns the peak nearest the supplied mass.
}
