% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspec.R
\name{subspec_spectra}
\alias{subspec_spectra}
\title{Find a the average isotopologue spectrum considering all spectra in the specified range.}
\usage{
subspec_spectra(specs, maxppm = Inf, s.aroundmax = Inf, id.cols = NULL)
}
\arguments{
\item{specs}{data.table containing columns mz, subspec, align.rtuserange, i, s, ppm, rt, mz.target}

\item{maxppm}{the maximum ppm error for an observation to be used in calculating the correction}

\item{s.aroundmax}{the number of scans on either side of the maxima in the specified range to aggregate. The default, Inf, takes the whole specified range.}

\item{id.cols}{columns to carry through}
}
\value{
data.table containing each averaged spectra, including normalized intensity on a scan to scan basis
}
\description{
\code{subspec_spectra} for every subspec finds the mean and standard deviation of each observed isotopologue.
}
